requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
cffi==1.14.1
chardet==3.0.4
Cybernator==0.6.0
discord.py==1.3.4
idna==2.10
multidict==4.7.6
numpy==1.19.1
psycopg2-binary==2.8.5
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.14.0
six==1.15.0
SQLAlchemy==1.3.18
websockets==8.1
yarl==1.5.0
youtube-dl==2020.6.16.1